migrate images feature
